# api/api.yml
...
openapi: "3.0.3"
info:
  version: 1.0.0
  title: Argos Service
  license:
    name: GPL v3
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
...
  /nodes/{parentId}/managementnodes:
    post:
      summary: create a management node
      operationId: createManagementNode
      tags:
        - management node
      parameters:
        - name: parentId
          in: path
          required: true
          description: this will be the parent id
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagementNode"
      responses:
        '201':
          description: created
          headers:
            Location:
              schema:
                type: string
                format: uri
              description: location where the created management node can be found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementNode"
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
    get:
      summary: get all management nodes for a parent
      operationId: getManagementNodes
      tags:
        - management node
      parameters:
        - name: parentId
          in: path
          required: true
          description: this is be the parent id
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '200':
          description: list of management nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManagementNode"
        '500':
          $ref: '#/components/responses/Error'
  /managementnodes/{managementNodeId}:
    get:
      summary: get a management node
      operationId: getManagementNode
      tags:
        - management node
      security:
        - bearerAuth: []
      parameters:
        - name: managementNodeId
          in: path
          required: true
          description: this will be the management node id
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementNode"
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
    delete:
      summary: delete management node by id.
      operationId: deleteManagementNodeById
      tags:
        - management node
      parameters:
        - name: managementNodeId
          in: path
          description: management node id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '204':
          description: management node deleted
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
  /nodes/{parentId}/projects:
    post:
      summary: create a project
      operationId: createProject
      tags:
        - project
      parameters:
        - name: parentId
          in: path
          required: true
          description: this will be the parent id
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Project"
      responses:
        '201':
          description: created
          headers:
            Location:
              schema:
                type: string
                format: uri
              description: location where the created project can be found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
    get:
      summary: get all projects for an ancestor
      operationId: getProjects
      tags:
        - project
      parameters:
        - name: parentId
          in: path
          required: true
          description: this will be the parent id
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '200':
          description: list of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        '500':
          $ref: '#/components/responses/Error'
  /projects/{projectId}:
    get:
      summary: get a project
      operationId: getProject
      tags:
        - project
      security:
        - bearerAuth: []

      parameters:
        - name: projectId
          in: path
          required: true
          description: this will be the project id
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
    delete:
      summary: delete project by id.
      operationId: deleteProjectById
      tags:
        - project
      parameters:
        - name: projectId
          in: path
          description: project id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '204':
          description: project deleted
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  responses:
    ValidationError:
      description: validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    Error:
      description: server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    UUID:
      type: string
      format: uuid
...
    NameString:
      type: string
      minLength: 1
      maxLength: 63
      pattern: '^([a-z]|[a-z][a-z0-9-]*[a-z0-9])?$'
...
    ManagementNode:
      type: object
      required:
        - name
        - parentId
        - pathToRoot
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          $ref: "#/components/schemas/NameString"
        pathToRoot:
          type: array
          maxItems: 256
          items:
            $ref: "#/components/schemas/UUID"
        parentId:
          $ref: "#/components/schemas/UUID"
    Project:
      type: object
      required:
        - name
        - parentId
        - pathToRoot
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          $ref: "#/components/schemas/NameString"
        pathToRoot:
          type: array
          maxItems: 256
          items:
            $ref: "#/components/schemas/UUID"
        parentId:
          $ref: "#/components/schemas/UUID"
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    ValidationError:
      type: object
      required:
        - messages
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ValidationMessage"
    ValidationMessage:
      type: object
      required:
        - message
        - type
      properties:
        field:
          type: string
        type:
          enum:
            - DATA_INPUT
            - MODEL_CONSISTENCY
        message:
          type: string
...